touchup

+++ .bin/ipfs-add-distfiles.sh
@@ -9,16 +9,25 @@ trap 'rm -r "${distdir_tmp}"' EXIT
 DISTDIR_TMP="${DISTDIR_TMP:-${distdir_tmp}}"
 DISTDIR="${DISTDIR:-/usr/portage/distfiles}"
 
-DISTDIR="${DISTDIR_TMP}" emerge --quiet --fetchonly "${1}" >> /dev/null || exit $?
+packages="$@"
+
+DISTDIR="${DISTDIR_TMP}" emerge --fetchonly "${packages}" || exit $?
+
+added_files=()
 
 for file in "${DISTDIR_TMP}"/* ; do
-	filename=$(basename ${file})
+  [[ -f "${file}" ]] || continue
+
+  filename=$(basename ${file})
   hash=$(ipfs add "${file}" --quiet --progress) || continue
 
   cd "${DISTDIR}"
   target_file="${DISTDIR}/${filename}"
+
   rm -f "${filename}"
   ln -s "/ipfs/${hash}" "${filename}"
   git add "${filename}"
-  git commit "${filename}" -m "link ${filename}"
+  added_files+="${filename}"
 done
+
+git commit "${added_files}" -m "link ${packages}"
